==> Austin/01_Trimming_Data_Frame_1.ipynb <==
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)

df=pd.read_csv('accepted_2007_to_2018Q4.csv',index_col='id')

def missingValuesInfo(df):
    total = df.isnull().sum().sort_values(ascending = False)
    percent = round(df.isnull().sum().sort_values(ascending = False)/len(df)*100, 2)
    temp = pd.concat([total, percent], axis = 1,keys= ['Total', 'Percent'])
    return temp.loc[(temp['Total'] > 0)]

missing_array=missingValuesInfo(df)

missing_array

df[(df.settlement_status!='ACTIVE') | (~df.settlement_status.isnull())].shape[0]

delete_column_list=list(missing_array.index[missing_array['Percent']>95])

lean_df=df.drop(delete_column_list,axis=1)

leaner_df=lean_df[~(lean_df.isnull().sum(axis=1)/lean_df.shape[1]>0.95)]

leanest_df=leaner_df.drop(['url','desc','policy_code','pymnt_plan','title'],axis=1)

leanest_df.to_csv('lean_df.csv')
