==> Austin/03_Imputing_Data_Frame.ipynb <==
#### Module Importation

import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

%reload_ext autoreload
%autoreload 
from HelperFunctions import missingValuesInfo

from sklearn.impute import SimpleImputer
from fancyimpute import KNN
from sklearn.model_selection import cross_val_score
from sklearn.ensemble import RandomForestRegressor, RandomForestClassifier

#### Helper Function

def imputer_score(imputer,estimator,x_miss,y):
    x_imputed=imputer.fit_transform(x_miss)
    impute_scores=cross_val_score(estimator,x_imputed,y,scoring='neg_mean_squared_error',cv=5)
    return -1*impute_scores.mean()

# Import Data

df=pd.read_csv('lean_df_5.csv',index_col='id')

# Variable Type Definition

df_number = df.select_dtypes(include = 'number')
df_object = df.select_dtypes(include = 'object')
df_category = df.select_dtypes(include = 'category')
df_boolean = df.select_dtypes(include = 'bool')
df_datetime = df.select_dtypes(include = 'datetime')
df_timedelta = df.select_dtypes(include = 'timedelta')
#######################################################
nominal_var=list(df_object.columns)
ordinal_var=list(df_number.columns)
continuous_var=list(df_number.columns)
time_var=list(df_datetime.columns)

# Instantiate Imputer 

imp_mean = SimpleImputer(missing_values=np.nan, strategy='mean')
imp_median = SimpleImputer(missing_values=np.nan, strategy= 'median')
imp_mode = SimpleImputer(missing_values=np.nan,strategy="most_frequent")
imp_const = SimpleImputer(missing_values=np.nan,fill_value='none',strategy="constant")
imp_const_n1 = SimpleImputer(missing_values=np.nan,fill_value=-999,strategy="constant")
imp_KNN = KNN(k=3)
# imp_MICE = IterativeImputer(random_state=0)


missingdf=missingValuesInfo(df)
missingdf

# Impute variables where missing may mean 0 with zero.
# impute non existent observation with -999

missinglist=list(missingdf.index)
impute_with_zero_list=['pub_rec_bankruptcies','tax_liens',\
             'delinq_amnt','total_acc','derog_records',\
             'delinq_2yrs','collections_12_mths_ex_med',\
             'chargeoff_within_12_mths','pub_rec_bankruptcies','dti']
impute_with_neg999=list(set(missinglist)-set(impute_with_zero_list))

for item in impute_with_zero_list:
    df.loc[df.loc[:,item].isnull(),item]=0

for item in impute_with_neg999:
    df.loc[df.loc[:,item].isnull(),item]=-999

missingValuesInfo(df)

# Only look at completed loans

final_df=df[df['loan_status']!='Current']

final_df.to_csv('pre_downsample_df.csv')

final_df.loan_status.value_counts()

# Take out the Current loans as the holdout dataset

holdout_df=df[df['loan_status']=='Current']

holdout_df.head()

holdout_df.to_csv('holdout_df.csv')

# df[continuous_var]=imp_mean.fit_transform(df[continuous_var])
# df[ordinal_var]=imp_mean.fit_transform(df[ordinal_var])
# df[nominal_var]=imp_const_n1.fit_transform(df[nominal_var])
# df[continuous_var]=imp_KNN.fit_transform(df[continuous_var])
# df[continuous_var]=imp_MICE.fit_transform(df[continuous_var])

# x=df_number.drop('SalePrice',axis=1)
# y=df_number.SalePrice
# regressor=RandomForestRegressor(n_estimators=100,random_state=0)
# classifier=RandomForestClassifier(n_estimators=100,random_state=0)
# imputer_score(imp_MICE,regressor,x,y)

# from sklearn.metrics import accuracy_score, log_loss
# from sklearn.neighbors import KNeighborsClassifier
# from sklearn.svm import SVC, LinearSVC, NuSVC
# from sklearn.tree import DecisionTreeClassifier
# from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier, GradientBoostingClassifier
# from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
# from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis
# classifiers = [
#     KNeighborsClassifier(3),
#     SVC(kernel="rbf", C=0.025, probability=True),
#     NuSVC(probability=True),
#     DecisionTreeClassifier(),
#     RandomForestClassifier(),
#     AdaBoostClassifier(),
#     GradientBoostingClassifier()
#     ]
# for classifier in classifiers:
#     pipe = Pipeline(steps=[('preprocessor', preprocessor),
#                       ('classifier', classifier)])
#     pipe.fit(X_train, y_train)   
#     print(classifier)
#     print("model score: %.3f" % pipe.score(X_test, y_test))

# from sklearn.impute import MissingIndicator
# x=np.array([[np.nan, 1, 3],[4, 0, np.nan],[8, 1, 0]])
# indicator = MissingIndicator()
# indicator.fit_transform(x)
