==> Austin/05b_Data_Transformations_and_Processing_FOR_HOLDOUT.ipynb <==
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

df=pd.read_csv('holdout_df.csv',index_col='id')

df.head()

df.shape

# Downsize holdout dataset to 1000 rows

df=df.sample(n=1000)

df.shape

df.head()

# Converting earliest credit into numeric, referencing issue_d

df.earliest_cr_line=pd.to_datetime(df.earliest_cr_line)
df.issue_d=pd.to_datetime(df.issue_d)
df.earliest_cr_line=df.issue_d-df.earliest_cr_line

df.earliest_cr_line=df.earliest_cr_line.astype('str')
df.earliest_cr_line=df.earliest_cr_line.str.extract('(\d+)')
df.earliest_cr_line=df.earliest_cr_line.astype('int')
df.loc[:,'earliest_cr_line']=df.earliest_cr_line/30/12
df=df.rename(columns={'earliest_cr_line':'credit_age_years'})

df.head()

# Converting employment length into numeric 

df.loc[:,'emp_length']=df.loc[:,'emp_length'].str.extract('(\d)').values

df.head()

# Converting profit into annualized return
#### Profit = total_payment - recovery_collection_fee = payout
#### payout/funded_amnt = (1+r)^(loan duration in years)

df.profit=np.exp((np.log(df.profit/df.funded_amnt))*12/df.loan_duration)-1

df=df.rename(columns={'profit':'return_rate'})

df.to_csv('down_sized_holdout_df.csv')

df.head()
